requirements.txt
alita
graphql
graphql-server-core